% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsFromCols.R
\name{binsFromCols}
\alias{binsFromCols}
\title{Concatenate the column names from a logical data.frame}
\usage{
binsFromCols(df, collapse = "_")
}
\arguments{
\item{df}{The \code{data.frame}}

\item{collapse}{The separator to use between the character variables in the column names}
}
\value{
A character vector
}
\description{
This checks that a data.frame is all logical values, 
then concatenates the column names using the given separator.
}
\details{
This is useful for turning sets of logical vectors into concatentaed strings 
able to be used for assigning values to bins.
}
\examples{
t <- c(TRUE, FALSE)
x <- data.frame(Type1 = sample(t, 10, TRUE), Type2 = sample(t, 10, TRUE))
binsFromCols(x)

}
\author{
Steve Pederson
}

