% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCI.R
\name{getCI}
\alias{getCI}
\title{Get Confidence Intervals after the bootstrap procedure}
\usage{
getCI(x, alpha, combined = FALSE)
}
\arguments{
\item{x}{The results from \code{\link{bootMeans}}}

\item{alpha}{Return (1 - alpha) confidence intervals}

\item{combined}{Combine all Tissues within the sample}
}
\value{
Returns a \code{tbl_df} with lower, median and upper quantiles for the interval. 
If \code{combined = TRUE} the \code{tbl_df} will only contain one row.
If \code{combined = FALSE} the \code{tbl_df} will contain one row per tissue.
}
\description{
Get Confidence Intervals after the bootstrap procedure
}
\details{
Extract the confidence intervals from the provided samples across all tissues
}
\author{
Steve Pederson
}

