% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVennCounts.R
\name{getVennCounts}
\alias{getVennCounts}
\title{Prepare columns for plotting as a Venn Diagram}
\usage{
getVennCounts(data, grpNames, rmRexp)
}
\arguments{
\item{data}{A data.frame or matrix}

\item{grpNames}{Optional. The names of the groups to be plotted}

\item{rmRexp}{Optional. A regular expression to be removed from the group names}
}
\value{
Returns a list with named objects ready to be plotted using \code{draw.quad.venn}, \code{draw.quintuple.venn} and associated functions
}
\description{
Prepare columns for plotting as a Venn Diagram
}
\examples{
x <- data.frame(a = rbinom(20, 1, 0.5), b = rbinom(20, 1, 0.5))
getVennCounts(x)

}
\author{
Steve Pederson
}

